refactor(auth): rename auth provider to notifier and improve auth state management
Implement new AuthNotifier with better state handling and analytics tracking
Clean up pubspec.yaml merge conflicts and formatting
Update text styling in auth screens to use theme consistently
Remove unused provider registration in main.dart
