+++ lib/presentation/auth/screens/splash_screen.dart
@@ -196,16 +196,19 @@ class _SplashScreenState extends State<SplashScreen>
                     builder: (context, child) {
                       return Opacity(
                         opacity: _fadeAnimation.value,
-                        child: const Text(
+                        child: Text(
                           'Lokapandu',
-                          style: TextStyle(
-                            color: Colors.white,
-                            fontSize: 28,
-                            fontWeight: FontWeight.bold,
-                            shadows: [
-                              Shadow(blurRadius: 8.0, color: Colors.black45),
-                            ],
-                          ),
+                          style: Theme.of(context).textTheme.headlineLarge
+                              ?.copyWith(
+                                color: Colors.white,
+                                fontWeight: FontWeight.bold,
+                                shadows: [
+                                  Shadow(
+                                    blurRadius: 8.0,
+                                    color: Colors.black45,
+                                  ),
+                                ],
+                              ),
                         ),
                       );
                     },
@@ -243,23 +246,19 @@ class _SplashScreenState extends State<SplashScreen>
     return Column(
       mainAxisSize: MainAxisSize.min,
       children: [
-        const Text(
+        Text(
           'Selamat Datang',
-          style: TextStyle(
-            color: Colors.white,
-            fontSize: 32,
+          style: Theme.of(context).textTheme.headlineSmall?.copyWith(
+            color: Colors.white.withValues(alpha: 0.9),
             fontWeight: FontWeight.bold,
-            shadows: [Shadow(blurRadius: 8.0, color: Colors.black54)],
           ),
           textAlign: TextAlign.center,
         ),
         const SizedBox(height: 16),
         Text(
           'Jelajahi dunia dengan Lokapandu\nsebagai pemandu andalanmu',
-          style: TextStyle(
+          style: Theme.of(context).textTheme.bodyMedium?.copyWith(
             color: Colors.white.withValues(alpha: 0.9),
-            fontSize: 18,
-            height: 1.5,
             shadows: const [Shadow(blurRadius: 6.0, color: Colors.black54)],
           ),
           textAlign: TextAlign.center,
