+++ pubspec.yaml
@@ -28,9 +28,6 @@ dependencies:
   shimmer: ^3.0.0
   google_fonts: ^6.3.2
 
-  
-
-
   brick_offline_first_with_supabase: ^2.1.0
   brick_sqlite: ^4.1.0
   brick_supabase: ^2.1.0
@@ -43,16 +40,12 @@ dependencies:
   build_runner: ^2.5.4
   go_router: ^16.2.4
   google_maps_flutter: ^2.13.1
-<<<<<<< HEAD
+
   shared_preferences: ^2.5.3
-=======
   google_sign_in: ^7.2.0
   firebase_auth: ^6.1.0
   firebase_analytics: ^12.0.2
->>>>>>> team
   
-  
-
 dev_dependencies:
   flutter_test:
     sdk: flutter
